requirements.txt
pandas
matplotlib
requests
python-dateutil
numpy
jupyter